Fix typos and bugs
